feat: Propsを使ってForAllTableにデータを渡す
