+++ react_typescript/src/components/Top.tsx
@@ -1,15 +1,51 @@
-import { useState } from "react";
-import { ForAllTable } from "./ForAll/forAllTable";
-import { ForMentorsTable } from "./ForMentors/forMentorsTable";
-import { ForStudentsTable } from "./ForStudents/forStudentsTable";
+import { useEffect, useState } from "react";
 import { NewRegisterForm } from "./NewRegisterForm";
+import { data } from "../api/data";
+import { MentorDataType } from "../type/MentorDataType";
+import { StudentDataType } from "../type/StudentDataType";
+import { ForAllTable } from "./ForAll/forAllTable";
 
 type TabTypes = "all" | "onlyStudents" | "onlyMentors";
+
+const getData: () => Promise<(MentorDataType | StudentDataType)[]> = async () =>
+  data;
 
 const Top = () => {
   const [activeTab, setActiveTab] = useState<TabTypes>("all");
 
-  // console.log(activeTab);
+  // 全データ
+  const [userListData, setUserListData] = useState<
+    (MentorDataType | StudentDataType)[]
+  >([]);
+
+  const [studentsData, setStudentsData] = useState<StudentDataType[]>([]);
+  const [mentorsData, setMentorsData] = useState<MentorDataType[]>([]);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      setUserListData(await getData());
+    };
+    fetchData();
+  }, []);
+
+  useEffect(() => {
+    const students: StudentDataType[] = [];
+    const mentors: MentorDataType[] = [];
+
+    userListData.forEach((data) => {
+      if (data.role === "student") {
+        students.push(data as StudentDataType);
+      } else {
+        mentors.push(data as MentorDataType);
+      }
+    });
+
+    setStudentsData(students);
+    setMentorsData(mentors);
+  }, [userListData]);
+
+  console.log("データ表示");
+  console.log(studentsData);
 
   return (
     <>
@@ -20,9 +56,11 @@ const Top = () => {
           メンターのみ
         </button>
       </div>
-      {activeTab === "all" && <ForAllTable />}
-      {activeTab === "onlyStudents" && <ForStudentsTable />}
-      {activeTab === "onlyMentors" && <ForMentorsTable />}
+      {activeTab === "all" && (
+        <ForAllTable studentsData={studentsData} mentorsData={mentorsData} />
+      )}
+      {/* {activeTab === "onlyStudents" && <ForStudentsTable />} */}
+      {/* {activeTab === "onlyMentors" && <ForMentorsTable />} */}
       <div>
         <NewRegisterForm />
       </div>
